DoTrueTypeFont can specify where packing and distance should occur in the pipeline TTF has commented out experimental summed area table support Convert DoXXX now has own header Texture export revamped to take TextureImage's support for double input to allow export of procedural textures directly TextureAtlas now takes option to not convert the textures, also returns the names of the textures used in the texture atlas Imageimport added which loads an image into a TextureImage<double> edtaa3 now using double to complement rest of texture pipeline texture float export fixed and tested convertInputToTextureImage now replaces internal data with full TextureImage for easy passing around TextureImage passed around via shared_ptr instead of copying

+++ src/libs/export/edtaa3func.h
@@ -2,7 +2,7 @@
 #ifndef YOLK_CONVERT_EDTAA3FUNC_H_
 #define YOLK_CONVERT_EDTAA3FUNC_H_ 1
 
-typedef float edtaa3real;
+typedef double edtaa3real;
 
 void computegradient(const edtaa3real *img, int w, int h, edtaa3real *gx, edtaa3real *gy);
 void edtaa3(const edtaa3real *img, const edtaa3real *gx, const edtaa3real *gy, int w, int h, short *distx, short *disty, edtaa3real *dist);
